feat: basic Z-string decoding
The original plan of a two-stage decoder failed. Alphabet changes
and locks only persist for the _current_ Z-string being processed,
but the previous method attempted to resolve abbreviations within
a Z-string, constructing a set of Z-chars comprising the original
string _and_ its expanded abbreviations in a single array.

When this composite array was passed to `Ztoa` it wasn't able to
be decoded properly. The expanded abbreviations sometimes ended
with padding characters (i.e. a series of 5's) which would've
been harmless if that expanded abbreviation was processed alone,
but because it was now part of the array of original Z-chars,
the padding characters were interpreted as alphabet changes that
persisted after the expanded abbreviation was processed.

The fix is to consolidate the Z-string decoding into a single
pass which processes Z-strings independently. Abbreviations are
processed into strings without mutating the original Z-char
array.